Moved code from camera.cpp, view_settings.cpp to viewportEngine.cpp
